+++ programs/drift/src/math/matching.rs
@@ -26,7 +26,7 @@ pub fn is_maker_for_taker(
     if taker_order.post_only || !maker_order.is_resting_limit_order(slot)? {
         Ok(false)
     // can make if taker order isn't resting (market order or limit going through auction)
-    } else if !taker_order.is_resting_limit_order(slot)? {
+    } else if !taker_order.is_resting_limit_order(slot)? || maker_order.post_only {
         Ok(true)
     // otherwise the maker must be older than the taker order
     } else {