+++ CHANGELOG.md
@@ -14,6 +14,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking
 
+## [2.19.0] - 2023-03-01
+
+### Features
+
+- program: allow post only to match older taker limit orders ([#378](https://github.com/drift-labs/protocol-v2/pull/378)) 
+- ts-sdk: serum subscriber supports websockets ([#365](https://github.com/drift-labs/protocol-v2/pull/365))
+- program: max number of subaccounts to 3000
+- program: amm spread logic more consistent across market by using liquidity ratio rather than base asset amount for inventory spread scaling([#374](https://github.com/drift-labs/protocol-v2/pull/374))
+- program: add pyth1M/pyth1K as OracleSource ([#375](https://github.com/drift-labs/protocol-v2/pull/375))
+
+### Fixes
+
+### Breaking
+
 ## [2.18.0] - 2023-02-24
 
 ### Features